pyproject.toml
[tool.poetry]
name = "crewai-ai-researcher"
version = "0.1.0"
description = ""
authors = ["Glen <glenbirkbeck@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
crewai = "^0.16"
python-dotenv = "1.0.0"
langchain-community = "^0.0.22"
crewai-tools = "^0.0.12"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"